+++ components/ADempiere/FieldDefinition/mixin/mixnWithDisplayColumn.js
@@ -31,6 +31,63 @@ export default {
   name: 'MixinWithDisplayColumn',
 
   computed: {
+    value: {
+      get() {
+        const { columnName, containerUuid, inTable } = this.metadata
+        // table records values
+        if (inTable) {
+          // implement container manager row
+          if (this.containerManager && this.containerManager.getCell) {
+            const value = this.containerManager.getCell({
+              containerUuid,
+              rowIndex: this.metadata.rowIndex,
+              columnName
+            })
+            if (!isEmptyValue(value)) {
+              return value
+            }
+          }
+        }
+
+        return store.getters.getValueOfFieldOnContainer({
+          parentUuid: this.metadata.parentUuid,
+          containerUuid,
+          columnName
+        })
+      },
+      set(value) {
+        const { columnName, containerUuid, inTable } = this.metadata
+
+        // table records values
+        if (inTable) {
+          // implement container manager row
+          if (this.containerManager && this.containerManager.setCell) {
+            this.containerManager.setCell({
+              containerUuid,
+              rowIndex: this.metadata.rowIndex,
+              columnName,
+              value
+            })
+          }
+        }
+
+        store.commit('updateValueOfField', {
+          parentUuid: this.metadata.parentUuid,
+          containerUuid,
+          columnName,
+          value
+        })
+        // update element column name
+        if (!this.metadata.isSameColumnElement) {
+          store.commit('updateValueOfField', {
+            parentUuid: this.metadata.parentUuid,
+            containerUuid,
+            columnName: this.metadata.elementName,
+            value
+          })
+        }
+      }
+    },
     displayedValue: {
       get() {
         // if (isEmptyValue(this.value)) {
